Extract Threat to a separate model
